Bounding box

+++ Code/Application.cpp
@@ -109,7 +109,7 @@ bool Application::Init()
 	CloseConfig();
 
 	//Framerate calculations
-	cap_time = 1000 / max_fps;
+	cap_time = (uint32)(1000 / max_fps);
 	curr_frame_time.Start();
 
 	return ret;
@@ -154,7 +154,7 @@ void Application::FinishUpdate()
 	}
 	seconds_since_startup = startup_time.ReadSec();
 	avg_fps = float(frame_count) / seconds_since_startup;
-	curr_frame_ms = curr_frame_time.ReadMs();
+	curr_frame_ms = (uint32)curr_frame_time.ReadMs();
 	UpateMsGraph(curr_frame_ms);
 
 	//Cap fps
@@ -318,7 +318,7 @@ bool Application::LoadAppConfiguration(JSON_Object * app_obj)
 		application_name = json_object_get_string(app_obj, "application name");
 		organization_name = json_object_get_string(app_obj, "organization name");
 		cap_fps = json_object_get_boolean(app_obj, "cap fps");
-		max_fps = json_object_get_number(app_obj, "max fps");
+		max_fps = (float)json_object_get_number(app_obj, "max fps");
 	}
 	return true;
 }
